host/ha: remove special-case observe mode code
Observe mode now uses the same code as manage mode, because manage mode
has to support observe-only behavior anyways for unmanaged resources.
